KAM-78: [FE] Widget post card component [zn-zorro] Init `app-news-card` component Renamed `reactive-single-select` folder to `reactive-select`
