+++ src/app/pages/main/pages/home/widgets/news-card/index.ts
@@ -1,0 +1,1 @@
+export * from './news-card.module';
