+++ src/app/pages/main/pages/home/widgets/news-card/news-card.component.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'app-news-card',
+  templateUrl: './news-card.component.html',
+  styleUrls: ['./news-card.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class NewsCardComponent {}
