+++ src/app/pages/main/pages/home/widgets/news-card/news-card.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { NewsCardComponent } from './news-card.component';
+
+@NgModule({
+  imports: [],
+  exports: [NewsCardComponent],
+  declarations: [NewsCardComponent],
+})
+export class NewsCardModule {}
