+++ src/app/pages/main/pages/media/media.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { MediaComponent } from './media.component';
 import { MediaRoutingModule } from './media-routing.module';
 import { RequestWrapperModule } from '@components/request-wrapper';
-import { ReactiveSelectModule } from '@widgets/reactive-fields/reactive-single-select';
+import { ReactiveSelectModule } from '@widgets/reactive-fields/reactive-select';
 import { ButtonModule } from '@widgets/button';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NzFormModule } from 'ng-zorro-antd/form';
